requirements.txt
setuptools>=68.0.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
pgvector==0.2.5
openai==1.60.2
tqdm==4.66.2
python-slugify==8.0.4
rich==13.7.0 